requirements.txt
PySide6>=6.5.0
pandas>=2.0.0
openpyxl>=3.1.0
lxml>=4.9.0
pytesseract>=0.3.10
Pillow>=10.0.0
requests>=2.31.0
